requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
idna==3.10
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.95
ultralytics-thop==2.0.6
urllib3==2.2.3

apriltag~=0.0.16
pynput~=1.7.7
pyserial~=3.5